Support LP ordering feature
:Release Notes:
Support LP ordering feature

:Detailed Notes:
The ordering feature is disabled in webOS OSE.
However, the feature is needed for dependent features in SAM.

For instance, listLaunchPoints subscription has dependency on it.

This commit enabled ordering feature again
to fix a bug about listLaunchPoints.
Basic policy of order is alphabetical.

:Testing Performed:
unit-test : passed

:QA Notes:

Change-Id: I87b978efdccb7f04ebaabf63b4d3df101e4cd53d