Getting env variables for session information
:Release Notes:
Getting env variables for session information

:Detailed Notes:
SAM should know it is running on container or not.
In addition, 'displayId' information which SAM can use in the env.
The 'displayId' can be used to replace 'launch' paramter.

Those variables are passed by another services.
In order to get those env variables, SAM needs to call 'getenv' function.
This commit is a base commit for further modification.

:Testing Performed:
Local test passed

:Issues Addressed:
[PLAT-102103] Modify SAM to be aware sessions

Change-Id: Ic519b3942aabd7fc785c7463b5d6524a20a3fd87
Reviewed-on: http://gpro.lge.com/c/webosose/sam/+/270626
Tested-by: Sangwoo Kang <sangwoo82.kang@lge.com>
Reviewed-by: Sangwoo Kang <sangwoo82.kang@lge.com>
Reviewed-by: OSE Helper Bot <ose@lge.com>