Fix timeout for retrying killing running applications.
:Release Notes:
Prevent sending thousands of messages to kill an application because of
a wrong timeout.

:Detailed Notes:
The GLib function used to set up a timeout to wait for closing or
killing applications, g_timeout_add, accepts an "interval" parameter
in milliseconds.

But, when an application is closed, the close transition timeout is
set to 1 (1ms), when it was actually expected to wait 1 second per
attempt to retry closing/killing an application. Fix the timeout
(and define a constant to store the value).

:Testing Performed:
Force an application to close because of memory usage using memstay
and launching several apps. The attempts to close should happen every
second.

:QA Notes:
Verification of the described test should also consider a fix in WAM
side, as it is not properly processing kill requests coming from
memory manager.

:Issues Addressed:
[WRN-12866] [RDX_CRASH][webos] /usr/bin/WebAppMgr in n/a (libcbe.so + 0x313a304)

Change-Id: I59f05b0d399185f5b9cae20676ff43bdcced03d5
Reviewed-on: http://gpro.lge.com/c/webosose/sam/+/321872
Tested-by: Guruprasadkn <guruprasad.kn@lge.com>
Reviewed-by: Guruprasadkn <guruprasad.kn@lge.com>